Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-customError-expected.txt


FAIL [input]  The validity.customError must be true if the custom validity error message is not empty assert_true: The validity.customError attribute should be true, when control is disabled. expected true got false
PASS [input]  The validity.customError must be false if the custom validity error message is empty
FAIL [button]  The validity.customError must be true if the custom validity error message is not empty assert_true: The validity.customError attribute should be true, when control is disabled. expected true got false
PASS [button]  The validity.customError must be false if the custom validity error message is empty
FAIL [select]  The validity.customError must be true if the custom validity error message is not empty assert_true: The validity.customError attribute should be true, when control is disabled. expected true got false
PASS [select]  The validity.customError must be false if the custom validity error message is empty
FAIL [textarea]  The validity.customError must be true if the custom validity error message is not empty assert_true: The validity.customError attribute should be true, when control is disabled. expected true got false
PASS [textarea]  The validity.customError must be false if the custom validity error message is empty
